requirements.txt
regex
urllib3
requests
datetime
python-dotenv